moved messages from Project to ProjectCache
